Resolved add gaji (#16)
* setup pm.xml and application.properties

* add templates folder and static file

* file: add form-base(temporary).html

* feat: create user model, add fragments, and halaman utama

* feat: add user, role, and gaji model

* feat: add login features (In Memory)

* feat: add datetime access on footer

* feat: add datetime access on footer

* feat: resolve add gaji conflict

+++ sipayroll/src/main/java/apap/tugasakhir/sipayroll/repository/GajiDb.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GajiDb extends JpaRepository<GajiModel, Long> {
+public interface GajiDb extends JpaRepository<GajiModel, Integer> {
 }
